did all changes required
