+++ server-side/email-service/Dockerfile
@@ -13,4 +13,4 @@ COPY . .
 
 EXPOSE 4001
 
-CMD ["node", "index.js"]
+CMD ["node", "run.js"]
