+++ server-side/email-service/app/package.json
@@ -2,10 +2,10 @@
   "name": "emailservice",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "run.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "node run.js"
   },
   "keywords": [],
   "author": "",
