--- server-side/email-service/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "emailservice",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "async-retry": "^1.3.3",
-    "bottleneck": "^2.19.5",
-    "express": "^4.19.2",
-    "googleapis": "^140.0.1",
-    "kafka-node": "^5.0.0",
-    "nodemailer": "^6.9.14"
-  }
-}
